Split out Era admin rpc methods
