+++ src/Nethermind/Nethermind.JsonRpc/Modules/Admin/AdminRpcModule.cs
@@ -11,13 +11,11 @@ using Nethermind.Blockchain.FullPruning;
 using Nethermind.Config;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
-using Nethermind.Era1;
 using Nethermind.Network;
 using Nethermind.Network.Config;
 using Nethermind.Specs.ChainSpecStyle;
 using Nethermind.State;
 using Nethermind.Stats.Model;
-using Nethermind.Synchronization.FastSync;
 using Nethermind.JsonRpc.Modules.Subscribe;
 using System.Text.Json;
 
@@ -36,7 +34,6 @@ public class AdminRpcModule : IAdminRpcModule
     private readonly IVerifyTrieStarter _verifyTrieStarter;
     private readonly IStateReader _stateReader;
     private NodeInfo _nodeInfo = null!;
-    private readonly IAdminEraService _eraService;
     private readonly ITrustedNodesManager _trustedNodesManager;
     private readonly ISubscriptionManager _subscriptionManager;
 
@@ -48,7 +45,6 @@ public class AdminRpcModule : IAdminRpcModule
         IVerifyTrieStarter verifyTrieStarter,
         IStateReader stateReader,
         IEnode enode,
-        IAdminEraService eraService,
         string dataDir,
         ManualPruningTrigger pruningTrigger,
         ChainParameters parameters,
@@ -64,7 +60,6 @@ public class AdminRpcModule : IAdminRpcModule
         _verifyTrieStarter = verifyTrieStarter ?? throw new ArgumentNullException(nameof(verifyTrieStarter));
         _stateReader = stateReader ?? throw new ArgumentNullException(nameof(stateReader));
         _pruningTrigger = pruningTrigger;
-        _eraService = eraService;
         _parameters = parameters ?? throw new ArgumentNullException(nameof(parameters));
         _trustedNodesManager = trustedNodesManager ?? throw new ArgumentNullException(nameof(trustedNodesManager));
 
@@ -200,16 +195,6 @@ public class AdminRpcModule : IAdminRpcModule
     public ResultWrapper<PruningStatus> admin_prune()
     {
         return ResultWrapper<PruningStatus>.Success(_pruningTrigger.Trigger());
-    }
-
-    public Task<ResultWrapper<string>> admin_exportHistory(string destination, int start = 0, int end = 0)
-    {
-        return ResultWrapper<string>.Success(_eraService.ExportHistory(destination, start, end));
-    }
-
-    public Task<ResultWrapper<string>> admin_importHistory(string source, int start = 0, int end = 0, string? accumulatorFile = null)
-    {
-        return ResultWrapper<string>.Success(_eraService.ImportHistory(source, start, end, accumulatorFile));
     }
 
     public ResultWrapper<string> admin_verifyTrie(BlockParameter block)
