+++ src/Nethermind/Nethermind.JsonRpc/Nethermind.JsonRpc.csproj
@@ -13,7 +13,6 @@
     <ProjectReference Include="..\Nethermind.Consensus\Nethermind.Consensus.csproj" />
     <ProjectReference Include="..\Nethermind.Core\Nethermind.Core.csproj" />
     <ProjectReference Include="..\Nethermind.Crypto\Nethermind.Crypto.csproj" />
-    <ProjectReference Include="..\Nethermind.Era1\Nethermind.Era1.csproj" />
     <ProjectReference Include="..\Nethermind.Evm\Nethermind.Evm.csproj" />
     <ProjectReference Include="..\Nethermind.Facade\Nethermind.Facade.csproj" />
     <ProjectReference Include="..\Nethermind.Network\Nethermind.Network.csproj" />
